backend/requirements.txt
# Django core
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0

# Database
psycopg2-binary==2.9.10

# WebSocket
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# Celery & Redis
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1

# APIs externes
requests==2.32.3
allocine-seances==0.0.14
unidecode==1.3.8

# Auth
argon2-cffi==23.1.0

# Config
python-decouple==3.8

# Storage
Pillow==11.0.0

# Tests
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0

# Dev
django-extensions==3.2.3
